3lab/project/requirements.txt
fastapi
uvicorn
sqlalchemy
email-validator
pydantic[email]
passlib[bcrypt]
python-jose
python-dotenv
alembic
celery
redis
websockets
aiofiles
requests
colorama
dnspython
idna
anyio
h11
pip
rsa
pyasn1
amqp
vine
ecdsa
Mako
MarkupSafe
sniffio
typing_extensions
six
kombu
billiard
urllib3
click
python-dateutil
click-repl
click-plugins
greenlet
starlette
email_validator
pydantic_core
wcwidth
annotated-types
prompt_toolkit